Container Changes
I have changed the classes and properties of most of the elements on the page so that they better fit the styling of the netmatters homepage. This includes the Services Cards, the Footer, the Articles Cards, and the welcome section.

They have improved drastically but will need some minor tweaks later on.