Updated Buttons
I have updated the buttons partial to include hover styles for the buttons used in the Tooltips and on other parts of the page. I will need to add in the styles for the telecoms, security and dev training buttons when I discover what colours are used on the page.